Refactor: move task_service to inbound, add Dockerfile and Docker Compose
