+++ cmd/main.go
@@ -5,6 +5,7 @@ import (
 	"github.com/bohexists/task-manager-svc/config"
 	"github.com/bohexists/task-manager-svc/internal/adapters/db"
 	"github.com/bohexists/task-manager-svc/internal/adapters/grpc"
+	"github.com/bohexists/task-manager-svc/ports/inbound"
 )
 
 func main() {
@@ -20,7 +21,7 @@ func main() {
 	taskService := app.NewTaskService(taskRepo)
 
 	// Initialize grpc server
-	grpcServiceServer := grpc.NewTaskServiceServer(taskService)
+	grpcServiceServer := inbound.NewTaskServiceServer(taskService)
 	// Start grpc server
 	grpc.StartGRPCServer(grpcServiceServer)
