+++ docker/docker-compose.yml
@@ -1,0 +1,30 @@
+version: '3.8'
+
+services:
+  task-manager-svc:
+    build: .
+    container_name: task-manager
+    ports:
+      - "50051:50051"
+    environment:
+      DB_HOST: task-db
+      DB_PORT: 3306
+      DB_USER: root
+      DB_PASSWORD: secret
+      DB_NAME: task_db
+    depends_on:
+      - task-db
+
+  task-db:
+    image: mysql:8.0
+    container_name: task-db
+    environment:
+      MYSQL_ROOT_PASSWORD: secret
+      MYSQL_DATABASE: task_db
+    ports:
+      - "3306:3306"
+    volumes:
+      - db_data:/var/lib/mysql
+
+volumes:
+  db_data:
